cluster-api/requirements.txt
fastapi[all]==0.75.0
pydantic==1.9
pymongo[srv]==3.11.0
python-dotenv==0.19.2
pytest==7.0.1
dapr==1.13.0
python-keycloak==3.11.1
pyparsing==3.1.2
python-jose==3.3.0